set up jwt cookie to header middleware

+++ app/config.py
@@ -52,7 +52,7 @@ class Configs:
                 ENV_VARS.get("APP_SECRET"),
                 ENV_VARS.get("JWT_SIGNITURE"),
                 ENV_VARS.get("JWT_HS_ALG"),
-                int(ENV_VARS.get("JWT_EXP_TIME")),
+                float(ENV_VARS.get("JWT_EXP_TIME")),
                 ENV_VARS.get("JWT_ISS"),
                 ENV_VARS.get("JWT_AUD")
             )
